Improve #credit logic
An Acima's lease request takes some time to process so it can happen that
even though a payment has been captured on Solidus' side, it still hasn't
been on Acima's side. If we send a refund request before Acima has processed
it, it will return an error.

Until now we have relied on vendors to watch out for this, but with this
commit we implement logic to check the current capture status on Acima's
side and based on that either follow with #credit's logic or revert to
#void's logic.